+++ .agentplane/tasks/202605251929-JZ4VPD/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202605251929-JZ4VPD/optimize-branch-pr-pr-check-artifact-fallback",
   "created_at": "2026-05-25T19:30:36.177Z",
-  "diffstat_sha256": "sha256:6bf54eff1c6e6f4e491b0653c675da4e3aad07295fb744deedda9b2cb42580ea",
+  "diffstat_sha256": "sha256:f63c531ae37ecf73bcef34e1514f8f464960c11ca720bf6fc920b907632a7969",
   "last_verified_at": "2026-05-25T19:55:06.438Z",
   "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
   "schema_version": 1,
